Fixing posts/mine and /posts endpoint + restructuring frontend directory
