+++ backend/src/post/types/create-post.dto.ts
@@ -1,0 +1,4 @@
+export class CreatePostDto {
+  userState!: number;
+  userCity!: number;
+}
