+++ frontend/src/app.tsx
@@ -1,22 +1,22 @@
+import NewPostForm from '@/components/CreatePost/NewPostForm/NewPostForm';
 import Home from '@/components/Home/Home';
 import { NavMenu } from '@/components/NavMenu/NavMenu';
-import NewPostForm from '@/components/NewPostForm/NewPostForm';
 import { Route, Switch } from 'wouter';
-import Loader from './components/Loader/Loader';
-import Login from './components/Login/Login';
-import Logout from './components/Logout/Logout';
-import Post from './components/Post/Post';
-import PostsHistory from './components/PostsHistory/PostsHistory';
+import Login from './components/Auth/Login/Login';
+import Logout from './components/Auth/Logout/Logout';
+import Post from './components/Posts/Post/Post';
+import PostsHistory from './components/Posts/PostsHistory/PostsHistory';
 import Profile from './components/Profile/Profile';
 import Test from './components/Test/Test';
 import { ThemeProvider } from './components/ThemeProvider/ThemeProvider';
 import { ModeToggle } from './components/ThemeToggle/ThemeToggle';
+import Loader from './components/Utility/Loader/Loader';
 import faq from './components/faq/faq';
 import AuthProvider from './context/AuthContext';
 // import { ErrorBoundary } from "react-error-boundary";
 // import RequestResponse from './components/RequestResponse/RequestResponse';
-import CenterThis from './components/CenterThis/CenterThis';
-import FormWithMessage from './components/FormWithMessage/FormWithMessage';
+import FormWithMessage from './components/CreatePost/FormWithMessage/FormWithMessage';
+import CenterThis from './components/Utility/CenterThis/CenterThis';
 import { Toaster } from './components/ui/toaster';
 import ContentLoadingProvider from './context/ContentLoadingContext';
