+++ frontend/src/components/Auth/Login/Login.tsx
@@ -10,7 +10,7 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '../ui/card';
+} from '../../ui/card';
 
 export default function Login() {
   const { session } = useContext(AuthContext);
