+++ frontend/src/components/CreatePost/NewPostForm/NewPostFormDropZone.tsx
