+++ frontend/src/components/Posts/PostsHistory/PostsHistory.tsx
