+++ frontend/src/context/AuthContext.tsx
@@ -1,7 +1,8 @@
 import { useToast } from '@/components/ui/use-toast';
 import { __supabase__ } from '@/constants';
 import { Session } from '@supabase/supabase-js';
-import { createContext, useEffect, useState } from 'react';
+import { createContext, useContext, useEffect, useState } from 'react';
+import { ContentLoadingContext } from './ContentLoadingContext';
 
 interface AuthContextValues {
   session: Session | null;
