+++ frontend/src/types/README.md
@@ -1,0 +1,1 @@
+Use this directory to maybe store zod schemas for DTOs
